chore: Changed the data type for currency rate from float to numeric

+++ chart/database/schemas/data.sql
@@ -1,7 +1,7 @@
 CREATE TABLE if not exists currency(
     from_currency TEXT NOT NULL,
     conv_currency TEXT NOT NULL,
-    rate FLOAT NOT NULL,
+    rate NUMERIC(30) NOT NULL,
     date DATE NOT NULL
 );
